test: Extract BuildLookAlike to a reusable module
